requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioschedule==0.5.2
aiosignal==1.3.2
annotated-types==0.7.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
asyncpg-lite==0.3.1.3
attrs==25.3.0
certifi==2025.1.31
frozenlist==1.5.0
greenlet==3.2.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
redis==5.2.1
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
yarl==1.19.0